Refactor 1

+++ ksaivulkan/Vulkan/Pipeline/VulkanShaderModule.cpp
@@ -1,14 +1,15 @@
 #include "VulkanShaderModule.hpp"
 
-ksai::VulkanShaderModule::VulkanShaderModule(const VulkanDevice& inDevice, const std::vector<uint32_t>& inSPIRV) 
-	: mDevice(inDevice.GetDeviceHandle())
-{
-    mCompiler = std::make_unique<spirv_cross::Compiler>(inSPIRV);
-    vk::ShaderModuleCreateInfo info(vk::ShaderModuleCreateFlags(), inSPIRV);
-    mModule = mDevice.createShaderModule(info);
-    mResources = mCompiler->get_shader_resources();
+ksai::VulkanShaderModule::VulkanShaderModule(const VulkanDevice& inDevice, const std::vector<uint32_t>& inSPIRV, bool inShouldDestroy) : mDevice(inDevice.GetDeviceHandle()) {
+               mCompiler = std::make_shared<spirv_cross::Compiler>(inSPIRV);
+               vk::ShaderModuleCreateInfo info(vk::ShaderModuleCreateFlags(), inSPIRV);
+               mResources     = mCompiler->get_shader_resources();
+               mModule        = mDevice.createShaderModule(info);
+               mShouldDestroy = inShouldDestroy;
 }
 
 ksai::VulkanShaderModule::~VulkanShaderModule() {
-	mDevice.destroyShaderModule(mModule);
+               if (mShouldDestroy) {
+                              mDevice.destroyShaderModule(mModule);
+               }
 }
